Final version of the GUI
This is the final version of the app. We decided to give up on having two separate windows for sending and receiving. Instead, the GUI is made of a single window which is split in two sections, one for sending and one for receiving. We also modified the settings and options available.